requirements.txt
kubernetes==26.1.0
ansible-navigator==3.5.0
pre-commit==3.6.0
ansible-creator
minio==7.2.3
